moved intervals to experimental
