+++ etna/experimental/prediction_intervals/__init__.py
@@ -1,0 +1,1 @@
+from etna.experimental.prediction_intervals.base import BasePredictionIntervals
